Moving to background/lib/posthog
- Changed `posthog.ts` to `background/lib`
- Updated import paths across services and UI items.
